changed slowprint import, Added slowprint to Ulitity from its own class file.
